Fix issue when asynchronous request on project list was not cancelled
When a project list was read from file, it did not cancel an ongoing asynchronous request.
The fix is to use cancellation token in this case.

Also:
1. Added protection from empty bodies to Discussion Box operations
2. Changed NamedCancellationTokenSource to TypedCancellationTokenSource
3. Fix minor UI controls inconsistency when a new merge request is selected.
Now all controls are disabled before loading a merge request.
They are enabled as soon as merge request is loaded if event it is a cancelled MR.