Fix issue with missed git repository update
Steps to reproduce the issue:
1. Select MR1 and launch diff tool => this sets CommitChecker of GitClient to MR1
2. Wait for a while... at the some moment a new version for MR2 appears in the GitLab.
3. Select MR2 and try to launch a diff tool
Result: Diff tool is not launched because we did not update git repository at the step 2.
And at the step 3 we don't update the repo because time stamps of periodic updates
are greater than timestamp of MR2's version from step 2.

Proposed solution:
1. In addition to CommitChecker add ProjectWatcher which notifies GitClientUpdater
about changes in a GitLab project. It is done by means of IProjectWatcher and it is
implemented by WorkflowUpdateChecker which checks GitLab by timer.
2. CommitChecker is still needed because:
a. WorkflowUpdateChecker does not check all GitLab hosts and projects but only selected ones.
b. WorkflowUpdateChecker checks things by timer and if a version just appeared we may miss it.
CommitChecker is used each time a user action (Launch Diff Tool or Show Discussions) occurs.
So CommitChecker is always bound to the selected MR.

Consequence:
Repositories which are not updated periodically by means of ProjectWatcher
will be updated on each user action by means of CommitChecker.

Also:
Fixed a bug with 'filter controls' which became disabled once filter is applied (due to zero MRs)
and never got enabled back.