Change auto-update implementation
GitLab server time is not synchronized with Client time and this caused issues
on auto-update because we compared timestamps in previous implementation.
Now we compare a list of Merge Requests from GitLab with a cached list of Merge Requests.
The same for Commits.

Also added ability to update UI when a closed merge request is detected.

This closes #58, #70.