Fix problems on upgrade
When upgraded from 0.5.2, host names updated on loading configuration but
did not save back to Config file immediately. This caused issues on attempt
to report a discussion because Config file did not have correct (with https
prefix) host names. Solution: When settings are changed on upgrade, save
them immediately to the Config file.
This closes #72