Fix a few UI bugs
1. Fixed crash on attempt to edit a destroyed textbox.
We no longer need to re-create textboxes because textbox height is adjusted on-the-fly, so problem code can be removed.
2. Fixed crash in Trace.TraceInformation call injected recently
3. Fixed crash on closing Discussions form with an editable TextBox
It was caused by access to a textbox that was disposed during async submit operation
4. Disabled all context menu actions on textboxes when these controls are removed by REFRESH (F5)
Unfortunately, there is no good way to do this, so just added checks to event handlers.
We had to disable them because to avoid working with `dangling` controls.
5. Added protection to refreshDiscussion() because it caused problems when an async operation is finished after REFRESH (F5).
For example, if asynchronous Delete returns the last discussion note, it tried to access a null Parent and crashed.
6. Removed early return which prevented subsequent git fetch calls
However, subsequent git fetch should not happen due to disabled UI controls

Closes #96