Store WorkflowState before async/await and modal dialog operations
This change is needed to allow the app to work with the state that it had when user made an action.
It was not important before because we did not have external calls and UI buttons that may
change state were grayed when needed. Now when code waits on async/await (or modal dialog)
an external call may occur which changes the state.

Also reorganized files a little bit.