Check if git is installed on start-up
The app expects that diff tool (BC3) and Git for Windows are installed properly and have
corresponding records in the system registry. If any of them is not installed, application cannot start.

To work with git, the app calls 'git' command which should be available via PATH environment variable.
To simplify code, the app just appends a folder with git binaries to the mrHelper process PATH environment variable.

Closes #147