Remove a resolved/unresolved thread from the cache
GitLab has a bug. It does not amend `updated_at` timestamp of notes when
they got resolved (or unresolved) by Resolve/Unresolve Thread button
or API call. This is a workaround.
We cannot do anything with Resolve Thread button but if a thread
is resolved from our application we can remove it from the cache
to reload completely on a next update request.
This solution does not solve a case when a thread is resolved from web
but it makes DiscussionManager much faster than in a version w/o cache.