Avoid missing notifications on Reload List
1. Reload List changed to Update List which just triggers a timer update
This simplifies Reload List logic significantly and allows processing
events from MergeRequestCache about new merge requests.

2. Added LoadedProjects event to IWorkflowEventNotifier.
This allows to avoid interleaving of timer updates and workflow events
that occured in MergeRequestCache (to reproduce it, AutoUpdatePeriodMs should be small).

3. Added protection from re-entrance of timer event handlers to UpdateManager.

4. Fixed logic of host switch in DiscussionManager.
Now DM is able to cancel requests for discussion updates for merge requests of all projects
that are not processed yet. For example, if we switch from Host A to Host B and Host A
has 10 projects, and at the moment of host switch 3 are processed and 7 are not,
we clear a queue that holds remaining 7 projects and for the three processed projects.
Cached discussions for 7 projects are cleared in onConnected().
If there is a MR being updated at the moment of host switch, it does not cache
anything due to _reconnect flag.

5. Simplified re-entrance protection in DiscussionManager.

6. TimeTrackingManager and DiscussionParser reset their state on host switch.