Make git helpers more friendly to host switch
In a case when host switched from A to B and back to A, git helpers
provided wrong data because on switch to B, they did not unsubscribe
from A projects and received project updates that were unable to handle
because merge request cache changed its internal state.
Then on change back to A, there were simply no git updates because
nothing changed from git point of view.

Proposed fix:
1. Unsubscribe from all A repositories on switch to B.
2. Force update on host switch to obtain initial snapshot.
This is excessive on application start-up because both "forced update"
within helper and a "silent git update" outside of it will trigger
git helper update procedure but it is the simplest fix for now.