Don't use --no-tags in old versions of git
Also fixed an issue with attempt to clone a repository when
UpdateContext has no SHA.
It may happen when diff tool is opened from Search tab and
project is missing in the list of projects, we need be able to clone it.
However in this case _latestFullUpdateTimestamp is not updated.

Closes #268