Improve Shallow Clone feature implementation
No matter if git uses shallow mode or not, the main app makes
two kinds of update requests:
- Full update request (contains a timestamp and full list of project SHA)
- Partial update request (contains a list of specific SHA)

Full update is requested when:
- project list is loaded on connect
- New MR or new commit in existing MR occurs

Partial update is requested when specific SHA need to be fetched.
This applies to UI requests and Git Helpers.

Internally, LocalGitRepositoryUpdater works this way.
Each update has two phases:
i. Either "git clone" or "git fetch" (if already cloned).
In Shallow Clone mode this step never calls "git fetch".
In non-Shallow Clone mode this step calls "git fetch" only
when we've got a full repository update request with a newer timestamp.

ii. Fetch a list of specific SHA one-by-one.
This allows to work with any commits (force pushed, squashed etc).
Since this may be a long loop, it is split in chunks of two SHA and can
be interleaved by partial update requests to in order to make UX better.

Apparently, this change in combination with restriction retirement
of support of repositories where single commit fetch is disabled
allows to retire CommitChainCreator completely.