Prevent ObjectDisposedException on exit
It is a non-functional change, just to make exit a bit better.
Since Process callbacks are invoked from another threads,
we may got in a situation when an invoke is already ready
to be done at the moment of MainForm Dispose.
Then on Dispose() a worker thread throws an exception.

Added a message pump call after we finalized all work but
before MainForm.Dispose() is called.