Fix splitter position loss on minimize/maximize
It is incorrect to call resetMinimumSizes() and updateMinimumSizes()
on window restore because these calls reset splitters to their default
positions. That calls were added earlier in window restore handling
in order to restore user-defined position of splitters when app launched
in minimized mode but it was a wrong decision.
Proposed fix is to restore user-defined splitter positions when the
main window is restored from minimized state first time.

Also fixed assertion caused by conflict between positions of two
radio buttons.

Closes #292