Fix commit order in diff tool
The intention is to provide the earliest revision as the left one for the diff tool
and the latest revision as the right one for the diff tool, no matter in what order
user selected items in the tree.

This did not work for commits which have the same timestamps because
ordering algorithm based on timestamps (Commit.Created_At) could not compare them.

As GitLab sends Commits in chronological order always and we show them in the
same order inside the Revision Browser, we can be sure that commits are shown
in the tree in chronological order.

Let's use tree item display index to select left and right revisions for the diff tool.

Closes #342