Check for updates when the main instance launches only
The problem was caused by second instance (started by Git Extensions) which
shown "Checking for updates" dialog but this check could not lead to showing
a link on a new version inside the application when user clicked Remind Later
because second instance cannot tell the main instance about found version.
So its the main instance responsibility to check for updates and install them.

Closes #384