Make DataCache.Disconnect() asynchronous to drop "connecting" state better
The problem manifested in the following scenario:
1) Click Search at Search tab
2) While Search is in progress, try to activate another search via custom protocol

Search #2 caused DataCacheConnectionCancelledException for Search #1.
DataCacheConnectionCancelledException handling in Search #1 caused
resetting DataCache state that has been already amended to Search #2.
So Search #2 was unexpectedly canceled.

So when a DataCache is being connected, and we want to cancel this procedure,
we have to wait for it.