Fix custom commands positions
repositionCustomCommands() depends on "Visible" state of controls and
should be called when Parent group box visibility changes.

Also fixed an issue with &= operator in enableCustomActions().
It is wrong to apply this operator to Visible state of a control
because this state may not represent actual visibility state
when a parent control is hidden.

Closes #434