Change list of special characters
It is hard to say why > and \ were included in a list special characters.
We can see now that they don't bring problems and only < character causes markdown issues.
Strings like "a<b" look bad.
Strings like "a > b" or "abc\" look ok.
However, the algorithm of DoesContainUnescapedSpecialCharacters() shall be changed
because now it treats good strings like "a<br>b" as suspicious ones.
"a<br>b" shall be treated as a good string.
For simplicity "a<foo>b" can be also treated as a good string because <foo>
does not corrupt a renedered markdown.