Fix two big problems
1. Fixed wrong behavior of "Select time tracking merge request" menu item.
The problem reproduced on attempt to "Select time tracking merge request"
being at the wrong host that time was currently tracking on.
The program could not switch host and select the MR.
Host switch code and Live/Recent/Search mode switch code was fixed and simplified.
Now ConnectionPage never initiates host switch, it is done from the MainForm only.
See comments in code.

2. If MRLV of one host had groups (was sorted by Project) and user switched
to MRLV of another host (which was not sorted by Project earlier), Configuration
settings were inherited (as we don't have per-host settings), but new Groups were
not created. On a next updateItemsWithGroups() call, Item.Group was accessed what
caused a crash. The same bug, but with different effects could be observed when
MRLV without groups was switched to MRLV which had groups before.
Now when MRLV settings change, an event is triggered and not Visible MRLVs
are able to create/destroy groups.